codes/requirements.txt
# ============================================================================
# DeepCode Python - Complete Requirements File
# All dependencies for 6 Featured Projects
# ============================================================================

# ============================================================================
# WEB FRAMEWORK & REST API
# ============================================================================
fastapi==0.104.1
uvicorn==0.24.0
starlette==0.27.0
pydantic==2.4.2
pydantic[email]==2.4.2

# ============================================================================
# DATABASE & ORM
# ============================================================================
sqlalchemy==2.0.23
psycopg2==2.9.9
alembic==1.12.1

# ============================================================================
# AUTHENTICATION & SECURITY
# ============================================================================
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
cryptography==41.0.7

# ============================================================================
# WEB SCRAPING
# ============================================================================
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0

# ============================================================================
# MACHINE LEARNING & DEEP LEARNING
# ============================================================================
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
matplotlib==3.8.2
opencv-python==4.8.1.78
pillow==10.0.0
scikit-learn==1.3.2
scikit-image==0.22.0

# ============================================================================
# NATURAL LANGUAGE PROCESSING
# ============================================================================
nltk==3.8.1
spacy==3.7.2
transformers==4.35.2
scikit-learn==1.3.2

# ============================================================================
# DEVOPS & CONTAINERIZATION
# ============================================================================
docker==7.0.0
kubernetes==28.1.0
pyyaml==6.0.1
httpx==0.25.1

# ============================================================================
# ASYNC & CONCURRENCY
# ============================================================================
asyncio==3.4.3
aiofiles==23.2.1

# ============================================================================
# UTILITIES & HELPERS
# ============================================================================
python-dotenv==1.0.0
click==8.1.7
colorama==0.4.6
tqdm==4.66.1

# ============================================================================
# DEVELOPMENT & TESTING
# ============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1
pylint==3.0.3

# ============================================================================
# DOCUMENTATION & CODE QUALITY
# ============================================================================
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# ============================================================================
# LOGGING & MONITORING
# ============================================================================
python-json-logger==2.0.7

# ============================================================================
# OPTIONAL: Performance & Optimization
# ============================================================================
# orjson==3.9.10          # Fast JSON serialization
# msgpack==1.0.7          # Efficient serialization
# redis==5.0.1            # Caching layer

# ============================================================================
# OPTIONAL: Advanced Features
# ============================================================================
# celery==5.3.4           # Task queue
# sqlparse==0.4.4         # SQL parsing
# dateutil==2.8.2         # Date utilities
